chore: mysql pg plugin move internal plugin
