+++ query/internal/plugins/builtin/plugins.go
@@ -1,1 +1,6 @@
 package builtin
+
+import (
+	_ "github.com/DataObserve/datav/query/internal/plugins/builtin/mysql"
+	_ "github.com/DataObserve/datav/query/internal/plugins/builtin/postgresql"
+)
