+++ query/internal/plugins/builtin/postgresql/postgresql.go
@@ -1,16 +1,18 @@
 package postgresql
 
 import (
+	"context"
 	"database/sql"
 	"database/sql/driver"
+	"fmt"
 	"reflect"
 	"sync"
 	"time"
 
-	pluginUtils "github.com/DataObserve/datav/query/internal/plugins/utils"
 	"github.com/DataObserve/datav/query/pkg/colorlog"
 	"github.com/DataObserve/datav/query/pkg/models"
 	"github.com/gin-gonic/gin"
+	_ "github.com/lib/pq"
 )
 
 var datasourceName = "postgresql"
@@ -27,7 +29,7 @@ func (*PostgreSQLPlugin) Query(c *gin.Context, ds *models.Datasource) models.Plu
 	query := c.Query("query")
 	conn, ok := conns[ds.Id]
 	if !ok {
-		conn, err := pluginUtils.ConnectToPostgreSQL(ds.URL, ds.Data["database"], ds.Data["username"], ds.Data["password"])
+		conn, err := connectToPostgreSQL(ds.URL, ds.Data["database"], ds.Data["username"], ds.Data["password"])
 		if err != nil {
 			colorlog.RootLogger.Warn("connect to postgresql error:", err, "ds_id", ds.Id, "url", ds.URL)
 			return models.PluginResult{
@@ -111,10 +113,41 @@ func (*PostgreSQLPlugin) Query(c *gin.Context, ds *models.Datasource) models.Plu
 }
 
 func (*PostgreSQLPlugin) TestDatasource(c *gin.Context) models.PluginResult {
-	return pluginUtils.TestPostgresqlDatasource(c)
+	url := c.Query("url")
+	database := c.Query("database")
+	username := c.Query("username")
+	password := c.Query("password")
+	db, err := connectToPostgreSQL(url, database, username, password)
+	if err != nil {
+		return models.GenPluginResult(models.PluginStatusError, err.Error(), nil)
+	}
+	if err = db.PingContext(context.Background()); err != nil {
+		return models.GenPluginResult(models.PluginStatusError, err.Error(), nil)
+	}
+	return models.GenPluginResult(models.PluginStatusSuccess, "", nil)
 }
 
 func init() {
 	// register datasource
 	models.RegisterPlugin(datasourceName, &PostgreSQLPlugin{})
+}
+
+func connectToPostgreSQL(url, database, username, password string) (*sql.DB, error) {
+	dsn := fmt.Sprintf("postgres://%s:%s@%s/%s", username, password, url, database)
+	colorlog.RootLogger.Debug("connect to postgresql dsn: ", dsn)
+
+	db, err := sql.Open("postgres", dsn)
+	if err != nil {
+		return nil, err
+	}
+	db.SetMaxIdleConns(5)
+	db.SetMaxOpenConns(5)
+	db.SetConnMaxLifetime(time.Duration(10) * time.Minute)
+	db.SetConnMaxIdleTime(time.Duration(10) * time.Minute)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*30)
+	defer cancel()
+	if err = db.PingContext(ctx); err != nil {
+		return nil, err
+	}
+	return db, nil
 }
