+++ query/internal/plugins/external/plugins.go
@@ -4,6 +4,4 @@ package external
 
 import (
 	_ "github.com/DataObserve/datav/query/internal/plugins/external/clickhouse"
-	_ "github.com/DataObserve/datav/query/internal/plugins/external/mysql"
-	_ "github.com/DataObserve/datav/query/internal/plugins/external/postgresql"
 )
